requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.1            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
django-extensions==1.8.1
django-rest-swagger==2.1.2
django==1.11.3
djangorestframework==3.6.3
flake8==3.3.0
idna==2.5                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.9.6             # via coreschema
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
mysqlclient==1.3.10
openapi-codec==1.3.1      # via django-rest-swagger
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pytz==2017.2              # via django
requests==2.18.1          # via coreapi
simplejson==3.11.1        # via django-rest-swagger
six==1.10.0               # via django-extensions
uritemplate==3.0.0        # via coreapi
urllib3==1.21.1           # via requests
